% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tiles_fixed_length.R
\name{create.tiles.fixed.length}
\alias{create.tiles.fixed.length}
\title{Create regions for testing methylation difference}
\usage{
create.tiles.fixed.length(data, tiles.length, common = NULL)
}
\arguments{
\item{data}{dataframe with specyfic columns: chr, poz, prob, no, meth, unmeth, meth.rate. This dataframe is result of function preprocessing.}

\item{tiles.length}{integer number that specifes maximum difference between minimum and maximum position in the same methylation regions}

\item{common}{logi value. If TRUE this function creates second regions group that are min postion is (min postion + max position)/2 of k-region and
max position is (min position + max position) of k+1 region.}
}
\value{
data.frame from parameter data with extra column tiles that is region id number within chromosomes and extra column
tiles.common if argument tiles.common is not null
}
\description{
The same regions are observation that maximum difference position is gaps.length argument
}
