% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_DMR.R
\name{find.DMR}
\alias{find.DMR}
\title{Finding DMR}
\usage{
find.DMR(data, methods, p.value.log.reg = NULL, p.value.reg.mixed = NULL,
  p.value.reg.corr.mixed = NULL)
}
\arguments{
\item{data}{There are two options:  1. dataframe with specyfic columns: chr, poz, prob, no, meth, unmeth, meth.rate.
This dataframe is result of function preprocessing.
2. dataframe with specyfic columns: chr, poz, prob, no, meth, unmeth, meth.rate, tiles and possible tiles.common columns. This dataframe is result of function create.tiles.min.gap or
create.tiles.fixed.length.}

\item{methods}{vectors with given methods. Possible values are: 'Wilcoxon', 'Ttest', 'KS', 'Reg.Log', 'Reg.Mixed',
'Reg.Corr.Mixed'.}

\item{p.value.log.reg}{if not NULL regions with p.value of prob variable smaller than p.value.log.reg are returned and  decreasingly ordered by absolute value of beta coefficient
of prob variable otherwise regions ale increasingly ordered by p.value}

\item{p.value.reg.mixed}{if not NULL regions with p.value of prob variable smaller than p.value.log.reg are returned and  decreasingly ordered by absolute value of beta coefficient
of prob variable otherwise regions ale increasingly ordered by p.value}

\item{p.value.reg.corr.mixed}{if not NULL regions with p.value of prob variable smaller than p.value.log.reg are returned and  decreasingly ordered by absolute value of beta coefficient
of prob variable otherwise regions ale increasingly ordered by p.value}
}
\value{
list object. Elements of list are results of given methods. The most interesting regions are on the top
}
\description{
Finding DMR by Wilcoxon, t-Student, Kolmogorov-Smirnow tests or logistic regression, logistic regression with mixed models,
logistic regression with mixed models with correlation matrix
}
