% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get.stats}
\alias{get.stats}
\title{Summarize regions}
\usage{
get.stats(data)
}
\arguments{
\item{data}{dataframe with specyfic columns: chr, poz, prob, no, meth, unmeth, meth.rate, tiles and possible tiles.common columns. This dataframe is result of function create.tiles.min.gap or
create.tiles.fixed.length}
}
\value{
data.frame which is summing-up regions specifed by tiles and tiles.common columns in data
}
\description{
Summarize regions by minimum, maximmum, mean, standard deviation of methylation rate in two probes and methylation diff rate withib two probes and estimated quantile basen od methylation diff rate
}
