% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocessing}
\alias{preprocessing}
\title{Preprocessing data}
\usage{
preprocessing(path.1, path.2)
}
\arguments{
\item{path.1}{a path to file.1 with methylation data}

\item{path.2}{a path to file.2 with methylation data}
}
\value{
data.frame with methylation values on common positions and chromosomes from data in path.1 and path.2
}
\description{
Preprocessing data for methylation analyses
}
